pyproject.toml
[project]
name = "unicon-runner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.9.2",
    "pathvalidate>=3.2.1",
    "jinja2>=3.1.4",
    "pika>=1.3.2",
    "python-dotenv>=1.0.1",
    "typer>=0.15.1",
    "rich>=13.9.4",
    "psutil>=6.1.1",
    "requests>=2.32.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "mypy>=1.11.2",
    "types-pika-ts>=1.3.0.20241203",
    "types-psutil>=6.1.0.20241102",
    "types-requests>=2.32.0.20241016",
]

[tool.ruff]
line-length = 100
lint.select = ["I", "F", "UP", "B", "SIM"]
